services/api-gateway/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.3
python-jose==3.3.0
python-multipart==0.0.9
starlette==0.36.3 